Refactor RandomGenerationStrategy and write tests for it
